requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
python-multipart>=0.0.6